Refactored the database query code to reduce redundancy
